fiexed some buggs

+++ include/spatialdomain.hpp
@@ -25,7 +25,7 @@ protected:
     cv::Mat m_sourceImg;
     int m_matType;
     std::string m_filepath;
-    int m_blockSize;
+    float m_blockSize;
 
     float m_H_blockiness;
     float m_V_blockiness;
